Move VariantPromotionRuleInfo to deiscount interface.py file
