+++ saleor/graphql/order/utils.py
@@ -5,8 +5,8 @@ from typing import TYPE_CHECKING, Dict, Iterable, List, Optional
 import graphene
 from django.core.exceptions import ValidationError
 
-from ...checkout.fetch import VariantPromotionRuleInfo
 from ...core.exceptions import InsufficientStock
+from ...discount.interface import VariantPromotionRuleInfo
 from ...order.error_codes import OrderErrorCode
 from ...order.utils import get_valid_shipping_methods_for_order
 from ...plugins.manager import PluginsManager
