+++ saleor/order/tests/test_order_utils.py
@@ -3,12 +3,9 @@ from decimal import Decimal
 import pytest
 from prices import Money, TaxedMoney
 
-from ...checkout.fetch import (
-    VariantPromotionRuleInfo,
-    fetch_checkout_info,
-    fetch_checkout_lines,
-)
+from ...checkout.fetch import fetch_checkout_info, fetch_checkout_lines
 from ...discount import DiscountType, DiscountValueType
+from ...discount.interface import VariantPromotionRuleInfo
 from ...giftcard import GiftCardEvents
 from ...giftcard.models import GiftCardEvent
 from ...graphql.order.utils import OrderLineData
