+++ saleor/tests/fixtures.py
@@ -38,11 +38,7 @@ from ..attribute.models import (
 )
 from ..attribute.utils import associate_attribute_values_to_instance
 from ..checkout import base_calculations
-from ..checkout.fetch import (
-    VariantPromotionRuleInfo,
-    fetch_checkout_info,
-    fetch_checkout_lines,
-)
+from ..checkout.fetch import fetch_checkout_info, fetch_checkout_lines
 from ..checkout.models import Checkout, CheckoutLine, CheckoutMetadata
 from ..checkout.utils import add_variant_to_checkout, add_voucher_to_checkout
 from ..core import EventDeliveryStatus, JobStatus
@@ -61,6 +57,7 @@ from ..discount import (
     RewardValueType,
     VoucherType,
 )
+from ..discount.interface import VariantPromotionRuleInfo
 from ..discount.models import (
     CheckoutLineDiscount,
     NotApplicable,
